recovery:updater: Changes to support 32->64 bit upgrades
We now support upgrading from 32 to 64 bit builds.
The check for device to package compatibility now
takes into account if the product name on the device
vs the product name on the upgrade build differs
only by the _32 or _64 trailing chars. In addition
we also force compile the updater binary as 32 bit
executable so that it can run on the 32 bit build
we are upgrading from.

Change-Id: Iebe6107e55719141fbcc496efa0c60ef1f9368c8

# Conflicts:
#	updater/install.cpp